Add GADT module and refactor FunctionDefinition structure
A new module, `gadt.rs`, is introduced. It is added to the public modules in `typing.rs`. The `Gadt` structure is removed from the `algebraic.rs` module. In parallel, the `FunctionDefinition` structure in the `functional.rs` module is refactored to handle GADT. This includes adding public visibility to some fields and creating a new `call` method.