Remove gadt.rs and update various typings
Removed the gadt.rs file as it is no longer needed. Updated the name of "SequenceType" to "Tuple" in multiple files for consistency. Added 'functional::FunctionType' as a public module in the typing.rs file. Added 'pub mod function' in value.rs. Made related minor changes in the README file.