Add parsing functionality and refactor 'typing' to 'types'
The parsing functionality was implemented, which enables raw binary data to be interpreted as different intrinsic types. Additionally, the "typing" module was refactored to "types". This better represents the module's purpose and increases code clarity. The addition of parsing functionality will make interoperability with raw byte data easier and more streamlined.